+++ tests/Other/Functions/ArgminTest.php
@@ -7,11 +7,6 @@ use PHPUnit\Framework\TestCase;
 
 class ArgminTest extends TestCase
 {
-    public function setUp()
-    {
-        //
-    }
-
     public function test_compute()
     {
         $value = Argmin::compute(['yes' => 0.8, 'no' => 0.2, 'maybe' => 0.0]);