+++ tests/Other/Functions/LogSumExpTest.php
@@ -7,11 +7,6 @@ use PHPUnit\Framework\TestCase;
 
 class LogSumExpTest extends TestCase
 {
-    public function setUp()
-    {
-        //
-    }
-
     public function test_compute()
     {
         $value = LogSumExp::compute([0.5, 0.4, 0.9, 1.0, 0.2, 0.9, 0.1, 0.5, 0.7]);